Move and update fake component spec definition
