+++ testing/integration/fake_test.go
@@ -8,72 +8,15 @@ package integration
 
 import (
 	"context"
-	"path/filepath"
-	"runtime"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/elastic/elastic-agent/pkg/component"
 	"github.com/elastic/elastic-agent/pkg/control/v2/client"
 	atesting "github.com/elastic/elastic-agent/pkg/testing"
 	"github.com/elastic/elastic-agent/pkg/testing/define"
 )
-
-var fakeComponent = atesting.UsableComponent{
-	Name:       "fake",
-	BinaryPath: mustAbs(filepath.Join("..", "..", "pkg", "component", "fake", "component", osExt("component"))),
-	Spec: &component.Spec{
-		Version: 2,
-		Inputs: []component.InputSpec{
-			{
-				Name:        "fake",
-				Description: "A fake input",
-				Platforms: []string{
-					"container/amd64",
-					"container/arm64",
-					"darwin/amd64",
-					"darwin/arm64",
-					"linux/amd64",
-					"linux/arm64",
-					"windows/amd64",
-				},
-				Shippers: []string{
-					"fake-shipper",
-				},
-				Command: &component.CommandSpec{},
-			},
-		},
-	},
-}
-
-var fakeShipper = atesting.UsableComponent{
-	Name:       "fake-shipper",
-	BinaryPath: mustAbs(filepath.Join("..", "..", "pkg", "component", "fake", "shipper", osExt("shipper"))),
-	Spec: &component.Spec{
-		Version: 2,
-		Shippers: []component.ShipperSpec{
-			{
-				Name:        "fake-shipper",
-				Description: "A fake shipper",
-				Platforms: []string{
-					"container/amd64",
-					"container/arm64",
-					"darwin/amd64",
-					"darwin/arm64",
-					"linux/amd64",
-					"linux/arm64",
-					"windows/amd64",
-				},
-				Outputs: []string{
-					"fake-action-output",
-				},
-				Command: &component.CommandSpec{},
-			},
-		},
-	},
-}
 
 var simpleConfig1 = `
 outputs:
@@ -160,19 +103,4 @@ func TestFakeComponent(t *testing.T) {
 		},
 	})
 	require.NoError(t, err)
-}
-
-func mustAbs(path string) string {
-	abs, err := filepath.Abs(path)
-	if err != nil {
-		panic(err)
-	}
-	return abs
-}
-
-func osExt(name string) string {
-	if runtime.GOOS == "windows" {
-		return name + ".exe"
-	}
-	return name
 }
